Add bookmark-list component
